fix: sign with safe and eoa correctly
